--- src/main/java/codes/kooper/blockify/events/PlayerExitStageEvent.java
@@ -1,48 +1,0 @@
-package codes.kooper.blockify.events;
-
-import codes.kooper.blockify.models.Stage;
-import lombok.Getter;
-import org.bukkit.entity.Player;
-import org.bukkit.event.Cancellable;
-import org.bukkit.event.Event;
-import org.bukkit.event.HandlerList;
-import org.jetbrains.annotations.NotNull;
-
-@Getter
-public class PlayerExitStageEvent extends Event implements Cancellable {
-    private boolean cancelled = false;
-    private static final HandlerList HANDLERS = new HandlerList();
-    private final Stage stage;
-    private final Player player;
-
-    /**
-     * Event that is called when a player exits a stage.
-     *
-     * @param stage  The stage the player exited.
-     * @param player The player that exited the stage.
-     */
-    public PlayerExitStageEvent(Stage stage, Player player) {
-        this.stage = stage;
-        this.player = player;
-    }
-
-    @Override
-    public @NotNull HandlerList getHandlers() {
-        return HANDLERS;
-    }
-
-    public static HandlerList getHandlerList() {
-        return HANDLERS;
-    }
-
-    @Override
-    public boolean isCancelled() {
-        return cancelled;
-    }
-
-    @Override
-    public void setCancelled(boolean cancel) {
-        this.cancelled = cancel;
-    }
-}
-
