+++ src/main/java/dev/twme/blocket/events/PlayerEnterStageEvent.java
@@ -1,0 +1,47 @@
+package dev.twme.blocket.events;
+
+import dev.twme.blocket.models.Stage;
+import lombok.Getter;
+import org.bukkit.entity.Player;
+import org.bukkit.event.Cancellable;
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.jetbrains.annotations.NotNull;
+
+@Getter
+public class PlayerEnterStageEvent extends Event implements Cancellable {
+    private boolean cancelled = false;
+    private static final HandlerList HANDLERS = new HandlerList();
+    private final Stage stage;
+    private final Player player;
+
+    /**
+     * Event that is called when a player enters a stage.
+     *
+     * @param stage  The stage the player entered.
+     * @param player The player that entered the stage.
+     */
+    public PlayerEnterStageEvent(Stage stage, Player player) {
+        this.stage = stage;
+        this.player = player;
+    }
+
+    @Override
+    public @NotNull HandlerList getHandlers() {
+        return HANDLERS;
+    }
+
+    public static HandlerList getHandlerList() {
+        return HANDLERS;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean cancel) {
+        this.cancelled = cancel;
+    }
+}
